proj4-starter/requirements.txt
django==2.2.5
djangorestframework==3.10.3